Binary cross entropy loss and adaptive error threshold at loss stage.

+++ datasets/oxford.py
@@ -112,7 +112,8 @@ class OxfordDataset(Dataset):
                                         self.config['cart_resolution'], self.config['cart_pixel_width'])  # 1 x H x W
         polar_mask = mean_intensity_mask(polar, self.mean_int_mask_mult)
         cart_mask = radar_polar_to_cartesian(azimuths, polar_mask, self.config['radar_resolution'],
-                                        self.config['cart_resolution'], self.config['cart_pixel_width'])
+                                             self.config['cart_resolution'],
+                                             self.config['cart_pixel_width']).astype(np.float32)
         # Get ground truth transform between this frame and the next
         time1 = int(self.frames[idx].split('.')[0])
         if idx + 1 < len(self.frames):
